refactor: extract model specs into internal/models to break circular import
Moves ModelSpec, KnownModels, ModelAliases, default model consts, DefaultMinScore,
and DimensionAwareMinScore from internal/embedder into a new internal/models leaf
package. internal/embedder re-exports all symbols for backward compatibility.
This breaks the future config → embedder → config circular import that would
arise when internal/config/service.go needs KnownModels and
internal/embedder/failover.go needs config.ConfigService.
